feat: add EntidadesCarousel component and integrate it into FeatureSection for enhanced UI consistency
